+++ packages/server/src/trpc/router/trpc-router.types.ts
@@ -35,9 +35,10 @@ export type InferTemplateMap<
     [key in keyof TMap]: TMap[key] extends CreateTemplateReturn<
       infer TProps,
       infer TMeta,
+      infer TDefaultMeta,
       infer TRes
     >
-      ? CreateTemplateProcedures<TProps, TMeta, TConfig, TRes>
+      ? CreateTemplateProcedures<TProps, TMeta, TDefaultMeta, TConfig, TRes>
       : TMap[key] extends AnyTemplateMap
       ? InferTemplateMap<TMap[key], TConfig>
       : never;
@@ -52,12 +53,13 @@ export type InferTemplateMap<
 export type CreateTemplateProcedures<
   TProps extends AnyTemplateProps,
   TMeta extends AnyMeta,
+  TDefaultMeta extends Partial<AnyMeta>,
   TConfig extends AnyRootConfig,
   TResponse extends any = void
 > = CreateRouterInner<
   TConfig,
   {
-    html: CreateTrpcQueryReturn<TProps, TMeta>;
+    html: CreateTrpcQueryReturn<TProps, TDefaultMeta>;
     send: CreateTrpcMutationReturn<TProps, TMeta, TResponse>;
   }
 >;