+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @brail/react
 
+## 1.0.0-beta.29
+
+### Patch Changes
+
+- Clarify RenderResult meta to defaultMeta and correct types, according to passed default meta value
+- Updated dependencies
+  - @brail/types@1.0.0-beta.21
+
 ## 1.0.0-beta.28
 
 ### Patch Changes